+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Fixed
 - Crash on invalid regular expressions in ``method``, ``endpoint`` and ``tag`` CLI options. `#403`_
 - Crash on non latin-1 encodable value in ``auth`` CLI option. `#404`_
 - Crash on invalid value in ``header`` CLI options. `#405`_
+- Crash on some invalid URLs in ``schema`` CLI option. `#406`_
 
 `0.24.1`_ - 2020-02-08
 ----------------------
@@ -694,6 +695,7 @@ Fixed
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0
 .. _0.2.0: https://github.com/kiwicom/schemathesis/compare/v0.1.0...v0.2.0
 
+.. _#406: https://github.com/kiwicom/schemathesis/issues/406
 .. _#405: https://github.com/kiwicom/schemathesis/issues/405
 .. _#404: https://github.com/kiwicom/schemathesis/issues/404
 .. _#403: https://github.com/kiwicom/schemathesis/issues/403